+++ src/LinearAlgebra/NBuffer.h
@@ -7,20 +7,20 @@
 template <typename T>
 class NBuffer {
  public:
-  NBuffer(int size) : size_(size) {
+  NBuffer(size_t size) : size_(size) {
     data_ = static_cast<T*>(std::aligned_alloc(32, size * sizeof(T)));
     std::memset(data_, 0, size * sizeof(T));
   }
 
   ~NBuffer() { std::free(data_); }
 
-  int size() const { return size_; }
+  size_t size() const { return size_; }
 
   T* data() const { return data_; }
 
-  T& operator[](int index) { return data_[index]; }
+  T& operator[](size_t index) { return data_[index]; }
 
-  const T& operator[](int index) const { return data_[index]; }
+  const T& operator[](size_t index) const { return data_[index]; }
 
   NBuffer clone() const {
     NBuffer<T> new_buffer(size_);
@@ -36,6 +36,6 @@ class NBuffer {
   void fill(T value) { std::fill(data_, data_ + size_, value); }
 
  private:
-  int size_;
+  size_t size_;
   T* data_;
 };
